Address API council comments - CTS
Test: all static shared lib CTS tests pass

bug:37327970

+++ hostsidetests/os/test-apps/StaticSharedLibConsumerApp1/src/android/os/lib/consumer1/UseSharedLibraryTest.java
@@ -101,7 +101,7 @@ public class UseSharedLibraryTest {
 
                 List<VersionedPackage> dependentPackages = sharedLib.getDependentPackages();
 
-                switch (sharedLib.getVersion()) {
+                switch ((int) sharedLib.getVersion()) {
                     case 1: {
                         firstLibFound = true;
                         assertSame(1, declaringPackage.getVersionCode());
